Bump ANDROID_PLATFORM to android-18 in CI/CQ
This is in preparation for a source change to require API level 18
(the NDK already requires 19). I've done this a bit differently from how
things are currently specified.

We currently parse the name of the builder in the "recipe":
https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipes/boringssl.py;l=214?q=boringssl.py&ss=chromium

This is a bit tedious to do, and I gather the preferred mechanism is for
the config file to set "properties". So I've set up a cmake_args
property which overrides the name-based defaults. Thus this CL sets the
properties in the config. It also mirrors some of the other configs
which we currently pick up by name. I'll follow this up with a CL to
remove the recipe options that this change makes redundant.

The other change is I've specified the API level with
ANDROID_PLATFORM=android-18, rather than ANDROID_NATIVE_API_LEVEL=18.
The latter comes from the old CMake toolchain file. The NDK's toolchain
file supports it for compatibility, but ANDROID_PLATFORM appears to be
the preferred name. It also supports both "18" and "android-18", but the
latter seems to be preferred from skimming the toolchain file.

Change-Id: I0b22a8feec534ab7efd453a291f3a26e10d1207c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/57486
Reviewed-by: Bob Beck <bbe@google.com>