Add two more Android FIPS builders
We support armv7 with FIPS in the shared library build, but this is
never tested in CI. We also support (though not I think for Android?)
delocate with aarch64. Since our only Linux aarch64 builders are
Android, add a statically linked Android builder.

Also update builders to use the properties added in
https://chromium-review.googlesource.com/c/chromium/tools/build/+/4296193,
which removes the last of the dependency on name-parsing.

Change-Id: Ic403c02a6fbd4ce65f4c3a24cf6cc6572686226a
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/57690
Reviewed-by: Adam Langley <agl@google.com>